mv implementation to dpdata/lmdb
